backend/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# 搜索引擎
elasticsearch==8.11.0

# OCR和图像处理
pytesseract==0.3.10
opencv-python==4.8.1.78
Pillow==10.1.0

# AI模型
tensorflow==2.15.0
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2

# 工具库
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.4

# 开发工具
pytest==7.4.3
black==23.11.0
flake8==6.1.0
